requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-allauth
django-rest-swagger
